requirements.txt
streamlit
openai
chromadb
python-dotenv
pandas
numpy